+++ csvkit/utilities/csvjoin.py
@@ -39,7 +39,8 @@ class CSVJoin(CSVKitUtility):
                  'Specify "0" to disable sniffing entirely, or "-1" to sniff the entire file.')
         self.argparser.add_argument(
             '-I', '--no-inference', dest='no_inference', action='store_true',
-            help='Disable type inference when parsing CSV input.')
+            help='Disable type inference (and --locale, --date-format, --datetime-format, --no-leading-zeroes) '
+                 'when parsing the input.')
 
     def main(self):
         if isatty(sys.stdin) and self.args.input_paths == ['-']: