+++ CHANGELOG.rst
@@ -7,6 +7,7 @@ Unreleased
 -  feat: :doc:`/scripts/csvsql` adds a :code:`--sql-delimiter` option, to set a different delimiter than ``;`` for the :code:`--query`, :code:`--before-insert` and :code:`after-insert` options.
 -  feat: :doc:`/scripts/sql2csv` adds a :code:`--execution-option` option.
 -  feat: :doc:`/scripts/sql2csv` uses the ``stream_results=True`` execution option, by default, to not load all data into memory at once.
+-  feat: Make :code:`csvcut -C` ignore unknown columns instead of throwing an exception.
 -  fix: :doc:`/scripts/csvsql` uses a default value of 1 for the :code:`--min-col-len` and :code:`--col-len-multiplier` options.
 -  fix: The :code:`--encoding` option defaults to the ``PYTHONIOENCODING`` environment variable if set.
 -  fix: For type inference, number takes priority over date-time, if not using the :code:`--datetime-format` option.